mac: Tweak tab spacing
- move the favicon 1px to the right
- reduce tab overlap by 1px
- reduce tab width by 6px
- move close button 5px to the left
- move the new tab button 1px to the right again
- reduce pinned tab width by 3px
- reduce distance from pinned tabs to remaining tabs by 1px

The Mac tab look now almost exactly matches CrOS. Remaining differences:
- text looks different (ok)
- tab overlap is 1px higher, for performance reasons (else, the
  loading spinner would overlap the tab on its left, causing a lot
  more painting during the spin animation)

BUG=52468,164448

Committed: https://src.chromium.org/viewvc/chrome?view=rev&revision=173053

Review URL: https://codereview.chromium.org/11580002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@173060 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b679a2c24b66adc24f4549245b70a400e3bbe0c7

+++ chrome/browser/ui/cocoa/tabs/tab_controller_unittest.mm
@@ -338,7 +338,7 @@ TEST_F(TabControllerTest, TitleViewLayout) {
                         [[controller titleView] frame]));
 
   // Resize the tab so that that the it grows.
-  tabFrame.size.width = [TabController maxTabWidth] * 0.75;
+  tabFrame.size.width = static_cast<int>([TabController maxTabWidth] * 0.75);
   [[controller view] setFrame:tabFrame];
 
   // The icon view and close button should be visible again and the title view
